+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Brilliant trait, makes piglins non-aggressive on armor. Requires the primary material to have the trait.
+- New trait condition `silentgear:primary_material`, requires the trait to be on the primary (first) material
 ### Fixed
 - GetMaterialStatsEvents not firing for graded materials (Silent's Gems supercharging)
 - Crude tool advancement not triggering